Add PaletteMapPriority
Traditionally libartpalette provided `PaletteGetSchedPriority' and
`PaletteSetSchedPriority`, which returned and took Java priority
arguments, thus attempting to entirely hide the mapping from clients.
This turned out to be impractical, because some ART components, e.g.
ART's system daemons, were tuned to use niceness values that did not
correspond to Java priorities, and thus used Posix/Linux `setpriority`
calls directly. This scheme also prevents us from easily caching
priorities set by `android.os.Process.setThreadPriority(int)`,
which traffics in Linux niceness values. And setThreadPriority()
clients should also really be able to inspect the Java priority
mapping.

Thus this adds `PaletteMapPriority()`, which is intended for use
in connection with Linux `setpriority(), and is preferred when
available. In the long term, this should replace the earlier calls.
(Whether or not we actually want to remove the earlier calls at some
point is deferred for now.)

In the short term, it is possible to emulate the new call with code
that sets each priority, and then uses Linux getpriority to examine
the result. This is clearly udesirable as a long term solution.

This is part of a larger effort by ag/32861626, and discussed in
go/thread-priorities-in-android to improve thread priority handling
in Android.

Note that this is being added at API level 36.1. This works because
libartpalette really uses dlsym to to determine whether a function is
present, and thus we rely on version checks only in the corresponding
test, in art. That check only works for integral API levels, but we
can easily work around this issue there.

This is not a public API, so there is no communication issue. ART
is the only client.

(Cherrypicked from aosp/3312350)

Flag: NONE Impractical See discussion on aosp/3312350 and aosp/3355710.
Bug: 389104950
Test: Builds and boots

Change-Id: Ib3f9c4f22b0375cece399074fd2c5fbf44ed29d5